rem-wm/requirements.txt
accelerate==1.7.0
aiofiles==23.2.1
altair==5.5.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
diffusers==0.32.2
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.1
gradio==3.50.2
gradio_client==0.6.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
intel-openmp==2021.4.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
loguru==0.6.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mkl==2021.4.0
mpmath==1.3.0
narwhals==1.41.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.6.0.66
orjson==3.10.18
packaging==25.0
pandas==2.2.3
peft==0.15.2
Pillow==9.2.0
pluggy==1.6.0
psutil==7.0.0
py==1.11.0
pydantic==1.10.2
pydub==0.25.1
pyparsing==3.2.3
pytest==7.1.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
safetensors==0.5.3
scikit-image==0.19.3
scipy==1.15.3
semantic-version==2.10.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.1
tbb==2021.11.0
tifffile==2025.5.10
timm==1.0.15
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.64.1
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==11.0.3
win32_setctime==1.2.0
zipp==3.22.0
